move stuff from Infra/Types to Infra/, move types around
