+++ src/Upcast/Infra/Resources/VPC.hs
@@ -16,9 +16,9 @@ import Data.Text (Text, unpack)
 
 import qualified Network.AWS.EC2 as EC2 -- (*)
 
-import Upcast.Infra.Types.Amazonka -- (*)
+import Upcast.Infra.Amazonka -- (*)
 import Upcast.Infra.NixTypes -- (*)
-import Upcast.Infra.Types.Common
+import Upcast.Infra.Types -- (*)
 
 -- *
